SPDB-87: build: fix clang compilation errors
This commit originally imported fuzzing tools from rocksdb.
However, after the rebase on 6.21 (from 6.11.4) only the compilation
fixes are left.

Original commit message:

Building them is a bit of a pain, but the OSS-Fuzz build scripts are
a good reference on how to do that:
https://github.com/google/oss-fuzz/tree/master/projects/rocksdb

This commit also fixes some clang compilation errors.

+++ db/compaction/compaction_picker_hybrid.cc
@@ -786,7 +786,8 @@ bool HybridCompactionPicker::NeedToRunLevelCompaction(
     return false;
   }
 
-  int forceCompactLevel = lastLevelInHyper - multiplier_[hyperLevelNum] - 6;
+  int forceCompactLevel =
+      int(lastLevelInHyper - multiplier_[hyperLevelNum]) - 6;
   size_t maxSize = sizeToCompact_[hyperLevelNum];
 
   if (hyperLevelNum == curNumOfHyperLevels_) {
@@ -1018,7 +1019,7 @@ bool HybridCompactionPicker::SelectNBuffers(
   }
 
   assert(lowest_level >= upper_level);
-  uint count = 0;
+  size_t count = 0;
   for (uint s = lowest_level; s >= upper_level; s--) {
     auto& levelFiles = vstorage->LevelFiles(s);
     if (!levelFiles.empty()) {
